+++ raster/src/test/scala/geotrellis/raster/CompositeTileSpec.scala
@@ -24,7 +24,7 @@ import org.scalatest.funspec.AnyFunSpec
 import spire.syntax.cfor._
 
 class CompositeTileSpec extends AnyFunSpec with TileBuilders with RasterMatchers with TestFiles {
-  describe("wrap") {
+  describe("CompositeTileSpec wrap") {
     it("wraps a literal raster") {
       val r =
         createTile(
@@ -49,7 +49,7 @@ class CompositeTileSpec extends AnyFunSpec with TileBuilders with RasterMatchers
         arr.toSet.size should be (1)
         values += arr(0)
       }
-      values.toSeq.sorted.toSeq should be (Seq(1, 2, 3, 4, 5, 6))
+      values.toSeq.sorted should be (Seq(1, 2, 3, 4, 5, 6))
 
       assertEqual(r, tiled)
     }
@@ -150,6 +150,27 @@ class CompositeTileSpec extends AnyFunSpec with TileBuilders with RasterMatchers
       actualExtent should be (extent)
       assertEqual(actualTile, tile)
 
+    }
+  }
+
+  describe("CompositeTile cellType combine") {
+    it("should union cellTypes") {
+      val int = {
+        val r =
+          createTile(
+            Array(1, 1, 1, 2, 2, 2, 3, 3, 3,
+              1, 1, 1, 2, 2, 2, 3, 3, 3,
+
+              4, 4, 4, 5, 5, 5, 6, 6, 6,
+              4, 4, 4, 5, 5, 5, 6, 6, 6),
+            9, 4)
+
+        val tl = TileLayout(3, 2, 3, 2)
+        CompositeTile.wrap(r, tl)
+      }
+      val dt = int.convert(DoubleCellType)
+
+      int.combine(dt)(_ + _).cellType shouldBe int.cellType.union(dt.cellType)
     }
   }
 }